Add PhysBoneLightController stub
