+++ com.liltoon.pcss-extension-1.4.9-final/Editor/VRCLightVolumesEditor.cs
@@ -2,15 +2,15 @@ using UnityEngine;
 using UnityEditor;
 using System.Linq;
 
-namespace lilToon.PCSS
+namespace lilToon.PCSS.Editor
 {
     /// <summary>
     /// VRC Light Volumes統合エディタ
     /// </summary>
-    [CustomEditor(typeof(VRCLightVolumesIntegration))]
-    public class VRCLightVolumesEditor : Editor
+    [CustomEditor(typeof(lilToon.PCSS.VRCLightVolumesIntegration))]
+    public class VRCLightVolumesEditor : UnityEditor.Editor
     {
-        private VRCLightVolumesIntegration integration;
+        private lilToon.PCSS.VRCLightVolumesIntegration integration;
         private bool showAdvancedSettings = false;
         private bool showStats = true;
