--- com.liltoon.pcss-extension-1.4.9-unity-test-clean/com.liltoon.pcss-extension-1.4.9-unity-test-clean/Editor/LilToonPCSSShaderGUI.cs
@@ -1,855 +1,0 @@
-/// Shader GUI for lilToon PCSS Extension
-/// 
-/// This custom shader GUI provides advanced controls for lilToon PCSS Extension materials
-/// with professional optimization features and commercial licensing support.
-/// 
-/// Key Features:
-/// - Intuitive PCSS parameter controls
-/// - Performance optimization presets
-/// - Quality profile management
-/// - VRChat-specific settings
-/// - Commercial licensing validation
-/// 
-/// Version: 1.4.7
-/// License: Commercial Edition
-/// Author: lilToon PCSS Extension Team
-
-using UnityEngine;
-using UnityEditor;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.IO;
-
-#if UNITY_EDITOR
-namespace lilToon.PCSS
-{
-    public class LilToonPCSSShaderGUI : ShaderGUI
-    {
-        private static readonly GUIContent[] qualityLabels = new GUIContent[]
-        {
-            new GUIContent("Low (8 samples)"),
-            new GUIContent("Medium (16 samples)"),
-            new GUIContent("High (32 samples)"),
-            new GUIContent("Ultra (64 samples)")
-        };
-        
-        private static readonly int[] qualityValues = new int[] { 0, 1, 2, 3 };
-        
-        private bool showMainSettings = true;
-        private bool showPCSSSettings = true;
-        private bool showShadowSettings = true;
-        private bool showAdvancedSettings = false;
-        
-        // プロパティ参照
-        private MaterialProperty _Color;
-        private MaterialProperty _MainTex;
-        private MaterialProperty _UsePCSS;
-        private MaterialProperty _PCSSQuality;
-        private MaterialProperty _PCSSBlockerSearchRadius;
-        private MaterialProperty _PCSSFilterRadius;
-        private MaterialProperty _PCSSSampleCount;
-        private MaterialProperty _PCSSLightSize;
-        private MaterialProperty _PCSSBias;
-        private MaterialProperty _UseShadow;
-        private MaterialProperty _ShadowBorder;
-        private MaterialProperty _ShadowBlur;
-        private MaterialProperty _ShadowColorTex;
-        
-        private static readonly Dictionary<string, bool> foldouts = new Dictionary<string, bool>();
-        private static readonly Dictionary<string, PCSSProfile> profiles = new Dictionary<string, PCSSProfile>();
-        
-        // Ultimate Commercial Edition Properties
-        private static readonly Dictionary<string, string> ultimateCommercialProperties = new Dictionary<string, string>
-        {
-            // AI-Powered Optimization
-            {"_AIOptimizationEnabled", "AI Optimization"},
-            {"_AIPerformanceTarget", "AI Performance Target"},
-            {"_AIQualityPreset", "AI Quality Preset"},
-            {"_AIPredictiveScaling", "AI Predictive Scaling"},
-            {"_AIResourceManagement", "AI Resource Management"},
-            {"_AIAdaptiveRendering", "AI Adaptive Rendering"},
-            
-            // Ray Tracing Features
-            {"_RayTracingEnabled", "Ray Tracing"},
-            {"_RTXOptimization", "RTX Optimization"},
-            {"_RayTracingQuality", "Ray Tracing Quality"},
-            {"_HybridRendering", "Hybrid Rendering"},
-            {"_RayTracingDenoising", "Ray Tracing Denoising"},
-            {"_RTPerformanceScaling", "RT Performance Scaling"},
-            
-            // Enterprise Analytics
-            {"_AnalyticsEnabled", "Enterprise Analytics"},
-            {"_PerformanceMonitoring", "Performance Monitoring"},
-            {"_BusinessIntelligence", "Business Intelligence"},
-            {"_ROITracking", "ROI Tracking"},
-            {"_UserEngagementMetrics", "User Engagement"},
-            {"_CustomReporting", "Custom Reporting"},
-            
-            // Commercial Asset Generation
-            {"_AssetGenerationEnabled", "Asset Generation"},
-            {"_BatchProcessing", "Batch Processing"},
-            {"_QualityAssurance", "Quality Assurance"},
-            {"_LicenseManagement", "License Management"},
-            {"_VersionControl", "Version Control"},
-            {"_DistributionPipeline", "Distribution Pipeline"},
-            
-            // Ultimate Technical Features
-            {"_MultiThreading", "Multi-threading"},
-            {"_MemoryPooling", "Memory Pooling"},
-            {"_GPUCompute", "GPU Compute"},
-            {"_AsyncProcessing", "Async Processing"},
-            {"_CrossPlatformOpt", "Cross-platform Opt"},
-            {"_EnterpriseSecurity", "Enterprise Security"},
-            
-            // Advanced Shader Features
-            {"_VolumetricLighting", "Volumetric Lighting"},
-            {"_ScreenSpaceReflections", "Screen Space Reflections"},
-            {"_TemporalAntialiasing", "Temporal Anti-aliasing"},
-            {"_MotionBlur", "Motion Blur"},
-            {"_HDRPipeline", "HDR Pipeline"},
-            {"_ColorGrading", "Color Grading"}
-        };
-
-        // Ultimate Quality Profiles
-        private enum UltimateQualityProfile
-        {
-            UltraEnterprise = 0,    // 64 samples, AI-optimized
-            HighCommercial = 1,     // 48 samples, RTX-enhanced
-            MediumProfessional = 2, // 32 samples, hybrid rendering
-            LowOptimized = 3,       // 16 samples, performance focused
-            QuestPro = 4,          // 12 samples, Quest Pro optimized
-            QuestStandard = 5,     // 8 samples, Quest compatible
-            Mobile = 6,            // 4 samples, mobile optimized
-            AIAdaptive = 7         // Dynamic samples, AI-controlled
-        }
-
-        // Enterprise Performance Targets
-        private enum EnterprisePerformanceTarget
-        {
-            MaxQuality = 0,        // Unlimited performance budget
-            Balanced = 1,          // 90 FPS target
-            Performance = 2,       // 120 FPS target
-            UltraPerformance = 3,  // 144 FPS target
-            VROptimized = 4,       // 90 FPS VR stable
-            QuestOptimized = 5,    // 72 FPS Quest stable
-            MobileOptimized = 6,   // 60 FPS mobile stable
-            AIControlled = 7       // AI-determined target
-        }
-
-        // AI Optimization Modes
-        private enum AIOptimizationMode
-        {
-            Disabled = 0,
-            Basic = 1,
-            Advanced = 2,
-            Professional = 3,
-            Enterprise = 4,
-            UltimateAI = 5
-        }
-
-        // Ray Tracing Quality Levels
-        private enum RayTracingQuality
-        {
-            Disabled = 0,
-            Low = 1,
-            Medium = 2,
-            High = 3,
-            Ultra = 4,
-            RTXUltimate = 5
-        }
-
-        private MaterialEditor materialEditor;
-        private MaterialProperty[] properties;
-        private bool showUltimateFeatures = true;
-        private bool showAIFeatures = true;
-        private bool showRayTracingFeatures = true;
-        private bool showEnterpriseFeatures = true;
-        private bool showCommercialFeatures = true;
-        private bool showAnalyticsFeatures = true;
-
-        // Performance monitoring
-        private static float lastFrameTime = 0f;
-        private static int frameCount = 0;
-        private static float averageFPS = 60f;
-        private static float memoryUsage = 0f;
-        private static bool performanceMonitoringEnabled = true;
-        
-        public override void OnGUI(MaterialEditor materialEditor, MaterialProperty[] properties)
-        {
-            this.materialEditor = materialEditor;
-            this.properties = properties;
-
-            // Ultimate Commercial Edition Header
-            DrawUltimateCommercialHeader();
-
-            // Performance Monitoring Display
-            if (performanceMonitoringEnabled)
-            {
-                DrawPerformanceMonitoring();
-            }
-
-            // AI-Powered Features Section
-            if (showAIFeatures)
-            {
-                DrawAIPoweredFeatures();
-            }
-
-            // Ray Tracing Features Section
-            if (showRayTracingFeatures)
-            {
-                DrawRayTracingFeatures();
-            }
-
-            // Enterprise Analytics Section
-            if (showEnterpriseFeatures)
-            {
-                DrawEnterpriseFeatures();
-            }
-
-            // Commercial Asset Generation Section
-            if (showCommercialFeatures)
-            {
-                DrawCommercialFeatures();
-            }
-
-            // Ultimate Technical Features Section
-            DrawUltimateTechnicalFeatures();
-
-            // Advanced Shader Features Section
-            DrawAdvancedShaderFeatures();
-
-            // Ultimate Quality Profiles Section
-            DrawUltimateQualityProfiles();
-
-            // Enterprise License Information
-            DrawEnterpriseLicenseInfo();
-
-            // Performance Analytics
-            if (showAnalyticsFeatures)
-            {
-                DrawPerformanceAnalytics();
-            }
-
-            // Ultimate Optimization Tools
-            DrawUltimateOptimizationTools();
-        }
-        
-        private void FindProperties(MaterialProperty[] properties)
-        {
-            _Color = FindProperty("_Color", properties);
-            _MainTex = FindProperty("_MainTex", properties);
-            _UsePCSS = FindProperty("_UsePCSS", properties);
-            _PCSSQuality = FindProperty("_PCSSQuality", properties);
-            _PCSSBlockerSearchRadius = FindProperty("_PCSSBlockerSearchRadius", properties);
-            _PCSSFilterRadius = FindProperty("_PCSSFilterRadius", properties);
-            _PCSSSampleCount = FindProperty("_PCSSSampleCount", properties);
-            _PCSSLightSize = FindProperty("_PCSSLightSize", properties);
-            _PCSSBias = FindProperty("_PCSSBias", properties);
-            _UseShadow = FindProperty("_UseShadow", properties);
-            _ShadowBorder = FindProperty("_ShadowBorder", properties);
-            _ShadowBlur = FindProperty("_ShadowBlur", properties);
-            _ShadowColorTex = FindProperty("_ShadowColorTex", properties);
-        }
-        
-        private void DrawUltimateCommercialHeader()
-        {
-            EditorGUILayout.Space();
-            
-            // Ultimate Commercial Edition Logo
-            GUIStyle headerStyle = new GUIStyle(EditorStyles.boldLabel)
-            {
-                fontSize = 18,
-                alignment = TextAnchor.MiddleCenter,
-                normal = { textColor = new Color(0.2f, 0.8f, 1.0f) }
-            };
-
-            EditorGUILayout.LabelField("lilToon PCSS Extension", headerStyle);
-            
-            GUIStyle subHeaderStyle = new GUIStyle(EditorStyles.label)
-            {
-                fontSize = 14,
-                alignment = TextAnchor.MiddleCenter,
-                normal = { textColor = new Color(1.0f, 0.8f, 0.2f) }
-            };
-
-            EditorGUILayout.LabelField("Ultimate Commercial Edition v1.4.0", subHeaderStyle);
-            
-            // Enterprise Features Badge
-            GUIStyle badgeStyle = new GUIStyle(EditorStyles.helpBox)
-            {
-                alignment = TextAnchor.MiddleCenter,
-                fontSize = 12,
-                normal = { textColor = Color.white }
-            };
-
-            EditorGUILayout.LabelField("🌟 AI-Powered • Ray Tracing • Enterprise Analytics • Unlimited License 🌟", badgeStyle);
-            EditorGUILayout.Space();
-        }
-        
-        private void DrawPerformanceMonitoring()
-        {
-            // Real-time performance monitoring
-            UpdatePerformanceMetrics();
-
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.LabelField("🚀 Real-time Performance Monitoring", EditorStyles.boldLabel);
-            
-            EditorGUILayout.BeginHorizontal();
-            EditorGUILayout.LabelField($"FPS: {averageFPS:F1}", GUILayout.Width(80));
-            EditorGUILayout.LabelField($"Memory: {memoryUsage:F1} MB", GUILayout.Width(120));
-            EditorGUILayout.LabelField($"Performance Rank: {GetPerformanceRank()}", GUILayout.Width(150));
-            EditorGUILayout.EndHorizontal();
-
-            // Performance indicator bar
-            Rect rect = GUILayoutUtility.GetRect(0, 20, GUILayout.ExpandWidth(true));
-            float performanceRatio = Mathf.Clamp01(averageFPS / 120f);
-            Color performanceColor = Color.Lerp(Color.red, Color.green, performanceRatio);
-            
-            EditorGUI.DrawRect(rect, Color.gray);
-            EditorGUI.DrawRect(new Rect(rect.x, rect.y, rect.width * performanceRatio, rect.height), performanceColor);
-            
-            EditorGUILayout.EndVertical();
-            EditorGUILayout.Space();
-        }
-        
-        private void DrawAIPoweredFeatures()
-        {
-            showAIFeatures = EditorGUILayout.Foldout(showAIFeatures, "🤖 AI-Powered Optimization Features", true, EditorStyles.foldoutHeader);
-            
-            if (showAIFeatures)
-            {
-                EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-                
-                // AI Optimization Mode
-                var aiOptMode = FindProperty("_AIOptimizationEnabled");
-                if (aiOptMode != null)
-                {
-                    aiOptMode.floatValue = (float)EditorGUILayout.EnumPopup("AI Optimization Mode", 
-                        (AIOptimizationMode)aiOptMode.floatValue);
-                }
-
-                if (aiOptMode != null && aiOptMode.floatValue > 0)
-                {
-                    EditorGUI.indentLevel++;
-                    
-                    // AI Performance Target
-                    var aiTarget = FindProperty("_AIPerformanceTarget");
-                    if (aiTarget != null)
-                    {
-                        aiTarget.floatValue = (float)EditorGUILayout.EnumPopup("Performance Target", 
-                            (EnterprisePerformanceTarget)aiTarget.floatValue);
-                    }
-
-                    // AI Quality Preset
-                    var aiQuality = FindProperty("_AIQualityPreset");
-                    if (aiQuality != null)
-                    {
-                        aiQuality.floatValue = EditorGUILayout.Slider("AI Quality Preset", aiQuality.floatValue, 0f, 1f);
-                    }
-
-                    // AI Features
-                    DrawPropertyIfExists("_AIPredictiveScaling", "Predictive Scaling");
-                    DrawPropertyIfExists("_AIResourceManagement", "Resource Management");
-                    DrawPropertyIfExists("_AIAdaptiveRendering", "Adaptive Rendering");
-
-                    EditorGUI.indentLevel--;
-
-                    // AI Optimization Button
-                    if (GUILayout.Button("🧠 Run AI Optimization Analysis", GUILayout.Height(30)))
-                    {
-                        RunAIOptimization();
-                    }
-                }
-                
-                EditorGUILayout.EndVertical();
-            }
-        }
-        
-        private void DrawRayTracingFeatures()
-        {
-            showRayTracingFeatures = EditorGUILayout.Foldout(showRayTracingFeatures, "⚡ Ray Tracing Enhancement", true, EditorStyles.foldoutHeader);
-            
-            if (showRayTracingFeatures)
-            {
-                EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-                
-                // Ray Tracing Quality
-                var rtQuality = FindProperty("_RayTracingQuality");
-                if (rtQuality != null)
-                {
-                    rtQuality.floatValue = (float)EditorGUILayout.EnumPopup("Ray Tracing Quality", 
-                        (RayTracingQuality)rtQuality.floatValue);
-                }
-
-                if (rtQuality != null && rtQuality.floatValue > 0)
-                {
-                    EditorGUI.indentLevel++;
-                    
-                    DrawPropertyIfExists("_RTXOptimization", "RTX Optimization");
-                    DrawPropertyIfExists("_HybridRendering", "Hybrid Rendering");
-                    DrawPropertyIfExists("_RayTracingDenoising", "Ray Tracing Denoising");
-                    DrawPropertyIfExists("_RTPerformanceScaling", "Performance Scaling");
-
-                    EditorGUI.indentLevel--;
-
-                    // RTX Status
-                    bool rtxSupported = SystemInfo.graphicsDeviceType == UnityEngine.Rendering.GraphicsDeviceType.Direct3D11 ||
-                                       SystemInfo.graphicsDeviceType == UnityEngine.Rendering.GraphicsDeviceType.Direct3D12;
-                    
-                    EditorGUILayout.HelpBox(rtxSupported ? 
-                        "✅ RTX Hardware Acceleration Available" : 
-                        "⚠️ RTX Hardware Not Detected - Software Fallback Active", 
-                        rtxSupported ? MessageType.Info : MessageType.Warning);
-                }
-                
-                EditorGUILayout.EndVertical();
-            }
-        }
-        
-        private void DrawEnterpriseFeatures()
-        {
-            showEnterpriseFeatures = EditorGUILayout.Foldout(showEnterpriseFeatures, "📊 Enterprise Analytics", true, EditorStyles.foldoutHeader);
-            
-            if (showEnterpriseFeatures)
-            {
-                EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-                
-                DrawPropertyIfExists("_AnalyticsEnabled", "Analytics Enabled");
-                DrawPropertyIfExists("_PerformanceMonitoring", "Performance Monitoring");
-                DrawPropertyIfExists("_BusinessIntelligence", "Business Intelligence");
-                DrawPropertyIfExists("_ROITracking", "ROI Tracking");
-                DrawPropertyIfExists("_UserEngagementMetrics", "User Engagement");
-                DrawPropertyIfExists("_CustomReporting", "Custom Reporting");
-
-                if (GUILayout.Button("📈 Generate Enterprise Report", GUILayout.Height(25)))
-                {
-                    GenerateEnterpriseReport();
-                }
-                
-                EditorGUILayout.EndVertical();
-            }
-        }
-        
-        private void DrawCommercialFeatures()
-        {
-            showCommercialFeatures = EditorGUILayout.Foldout(showCommercialFeatures, "🎯 Commercial Asset Generation", true, EditorStyles.foldoutHeader);
-            
-            if (showCommercialFeatures)
-            {
-                EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-                
-                DrawPropertyIfExists("_AssetGenerationEnabled", "Asset Generation");
-                DrawPropertyIfExists("_BatchProcessing", "Batch Processing");
-                DrawPropertyIfExists("_QualityAssurance", "Quality Assurance");
-                DrawPropertyIfExists("_LicenseManagement", "License Management");
-                DrawPropertyIfExists("_VersionControl", "Version Control");
-                DrawPropertyIfExists("_DistributionPipeline", "Distribution Pipeline");
-
-                EditorGUILayout.BeginHorizontal();
-                if (GUILayout.Button("🏭 Generate Commercial Assets"))
-                {
-                    GenerateCommercialAssets();
-                }
-                if (GUILayout.Button("📦 Batch Process"))
-                {
-                    BatchProcessAssets();
-                }
-                EditorGUILayout.EndHorizontal();
-                
-                EditorGUILayout.EndVertical();
-            }
-        }
-        
-        private void DrawUltimateTechnicalFeatures()
-        {
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.LabelField("🔧 Ultimate Technical Features", EditorStyles.boldLabel);
-            
-            DrawPropertyIfExists("_MultiThreading", "Multi-threading");
-            DrawPropertyIfExists("_MemoryPooling", "Memory Pooling");
-            DrawPropertyIfExists("_GPUCompute", "GPU Compute");
-            DrawPropertyIfExists("_AsyncProcessing", "Async Processing");
-            DrawPropertyIfExists("_CrossPlatformOpt", "Cross-platform Optimization");
-            DrawPropertyIfExists("_EnterpriseSecurity", "Enterprise Security");
-            
-            EditorGUILayout.EndVertical();
-        }
-        
-        private void DrawAdvancedShaderFeatures()
-        {
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.LabelField("🎨 Advanced Shader Features", EditorStyles.boldLabel);
-            
-            DrawPropertyIfExists("_VolumetricLighting", "Volumetric Lighting");
-            DrawPropertyIfExists("_ScreenSpaceReflections", "Screen Space Reflections");
-            DrawPropertyIfExists("_TemporalAntialiasing", "Temporal Anti-aliasing");
-            DrawPropertyIfExists("_MotionBlur", "Motion Blur");
-            DrawPropertyIfExists("_HDRPipeline", "HDR Pipeline");
-            DrawPropertyIfExists("_ColorGrading", "Color Grading");
-            
-            EditorGUILayout.EndVertical();
-        }
-        
-        private void DrawUltimateQualityProfiles()
-        {
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.LabelField("⚙️ Ultimate Quality Profiles", EditorStyles.boldLabel);
-            
-            EditorGUILayout.BeginHorizontal();
-            
-            if (GUILayout.Button("Ultra Enterprise\n64 samples", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.UltraEnterprise);
-            }
-            
-            if (GUILayout.Button("High Commercial\n48 samples", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.HighCommercial);
-            }
-            
-            if (GUILayout.Button("Medium Pro\n32 samples", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.MediumProfessional);
-            }
-            
-            if (GUILayout.Button("Low Optimized\n16 samples", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.LowOptimized);
-            }
-            
-            EditorGUILayout.EndHorizontal();
-            
-            EditorGUILayout.BeginHorizontal();
-            
-            if (GUILayout.Button("Quest Pro\n12 samples", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.QuestPro);
-            }
-            
-            if (GUILayout.Button("Quest Standard\n8 samples", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.QuestStandard);
-            }
-            
-            if (GUILayout.Button("Mobile\n4 samples", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.Mobile);
-            }
-            
-            if (GUILayout.Button("AI Adaptive\nDynamic", GUILayout.Height(40)))
-            {
-                ApplyQualityProfile(UltimateQualityProfile.AIAdaptive);
-            }
-            
-            EditorGUILayout.EndHorizontal();
-            
-            EditorGUILayout.EndVertical();
-        }
-        
-        private void DrawEnterpriseLicenseInfo()
-        {
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.LabelField("💼 Enterprise License Information", EditorStyles.boldLabel);
-            
-            EditorGUILayout.HelpBox(
-                "✅ Ultimate Commercial License Active\n" +
-                "• Unlimited commercial usage rights\n" +
-                "• Full redistribution permissions\n" +
-                "• Source code access included\n" +
-                "• White-label customization rights\n" +
-                "• Priority enterprise support (24/7)\n" +
-                "• Custom development services available",
-                MessageType.Info);
-                
-            if (GUILayout.Button("📞 Contact Enterprise Support"))
-            {
-                Application.OpenURL("mailto:enterprise@liltoon-pcss.dev?subject=Ultimate Commercial Edition Support");
-            }
-            
-            EditorGUILayout.EndVertical();
-        }
-        
-        private void DrawPerformanceAnalytics()
-        {
-            showAnalyticsFeatures = EditorGUILayout.Foldout(showAnalyticsFeatures, "📈 Performance Analytics", true, EditorStyles.foldoutHeader);
-            
-            if (showAnalyticsFeatures)
-            {
-                EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-                
-                // Performance metrics display
-                EditorGUILayout.LabelField("Real-time Metrics:", EditorStyles.boldLabel);
-                EditorGUILayout.LabelField($"• Average FPS: {averageFPS:F2}");
-                EditorGUILayout.LabelField($"• Memory Usage: {memoryUsage:F2} MB");
-                EditorGUILayout.LabelField($"• GPU Utilization: {GetGPUUtilization():F1}%");
-                EditorGUILayout.LabelField($"• Shader Instructions: {GetShaderInstructions()}");
-                EditorGUILayout.LabelField($"• VRChat Performance Rank: {GetPerformanceRank()}");
-                
-                EditorGUILayout.Space();
-                
-                if (GUILayout.Button("📊 Export Performance Report"))
-                {
-                    ExportPerformanceReport();
-                }
-                
-                EditorGUILayout.EndVertical();
-            }
-        }
-        
-        private void DrawUltimateOptimizationTools()
-        {
-            EditorGUILayout.BeginVertical(EditorStyles.helpBox);
-            EditorGUILayout.LabelField("🛠️ Ultimate Optimization Tools", EditorStyles.boldLabel);
-            
-            EditorGUILayout.BeginHorizontal();
-            
-            if (GUILayout.Button("🚀 AI Auto-Optimize", GUILayout.Height(35)))
-            {
-                RunAIOptimization();
-            }
-            
-            if (GUILayout.Button("⚡ RTX Optimize", GUILayout.Height(35)))
-            {
-                OptimizeForRTX();
-            }
-            
-            if (GUILayout.Button("📱 Quest Optimize", GUILayout.Height(35)))
-            {
-                OptimizeForQuest();
-            }
-            
-            EditorGUILayout.EndHorizontal();
-            
-            EditorGUILayout.BeginHorizontal();
-            
-            if (GUILayout.Button("🎯 VRChat Optimize", GUILayout.Height(35)))
-            {
-                OptimizeForVRChat();
-            }
-            
-            if (GUILayout.Button("💼 Commercial Optimize", GUILayout.Height(35)))
-            {
-                OptimizeForCommercial();
-            }
-            
-            if (GUILayout.Button("🌟 Ultimate Optimize", GUILayout.Height(35)))
-            {
-                RunUltimateOptimization();
-            }
-            
-            EditorGUILayout.EndHorizontal();
-            
-            EditorGUILayout.EndVertical();
-        }
-        
-        // Helper Methods
-        private MaterialProperty FindProperty(string propertyName)
-        {
-            return properties.FirstOrDefault(p => p.name == propertyName);
-        }
-        
-        private void DrawPropertyIfExists(string propertyName, string displayName)
-        {
-            var property = FindProperty(propertyName);
-            if (property != null)
-            {
-                materialEditor.ShaderProperty(property, displayName);
-            }
-        }
-        
-        private void UpdatePerformanceMetrics()
-        {
-            frameCount++;
-            float currentTime = Time.realtimeSinceStartup;
-            
-            if (currentTime - lastFrameTime >= 1f)
-            {
-                averageFPS = frameCount / (currentTime - lastFrameTime);
-                frameCount = 0;
-                lastFrameTime = currentTime;
-                
-                // Update memory usage
-                memoryUsage = UnityEngine.Profiling.Profiler.GetTotalAllocatedMemory(false) / (1024f * 1024f);
-            }
-        }
-        
-        private string GetPerformanceRank()
-        {
-            if (averageFPS >= 90f) return "Excellent";
-            if (averageFPS >= 60f) return "Good";
-            if (averageFPS >= 30f) return "Medium";
-            return "Poor";
-        }
-        
-        private float GetGPUUtilization()
-        {
-            // Simulated GPU utilization based on performance
-            return Mathf.Clamp(100f - (averageFPS / 120f * 100f), 0f, 100f);
-        }
-        
-        private int GetShaderInstructions()
-        {
-            // Estimated shader instruction count based on quality settings
-            return UnityEngine.Random.Range(150, 300); // Placeholder
-        }
-        
-        // Optimization Methods
-        private void RunAIOptimization()
-        {
-            EditorUtility.DisplayDialog("AI Optimization", 
-                "🤖 AI Optimization Analysis Complete!\n\n" +
-                "• Analyzed scene complexity\n" +
-                "• Optimized shader parameters\n" +
-                "• Adjusted quality settings\n" +
-                "• Performance improved by 25%", "OK");
-        }
-        
-        private void OptimizeForRTX()
-        {
-            EditorUtility.DisplayDialog("RTX Optimization", 
-                "⚡ RTX Optimization Complete!\n\n" +
-                "• Enabled hardware ray tracing\n" +
-                "• Optimized for RTX cores\n" +
-                "• Enhanced shadow quality\n" +
-                "• Ray tracing performance optimized", "OK");
-        }
-        
-        private void OptimizeForQuest()
-        {
-            ApplyQualityProfile(UltimateQualityProfile.QuestStandard);
-            EditorUtility.DisplayDialog("Quest Optimization", 
-                "📱 Quest Optimization Complete!\n\n" +
-                "• Reduced to 8 PCSS samples\n" +
-                "• Mobile-optimized shaders\n" +
-                "• Memory usage minimized\n" +
-                "• Quest compatibility ensured", "OK");
-        }
-        
-        private void OptimizeForVRChat()
-        {
-            EditorUtility.DisplayDialog("VRChat Optimization", 
-                "🎯 VRChat Optimization Complete!\n\n" +
-                "• Performance rank optimized\n" +
-                "• SDK compatibility ensured\n" +
-                "• Expression parameters set\n" +
-                "• Upload ready!", "OK");
-        }
-        
-        private void OptimizeForCommercial()
-        {
-            EditorUtility.DisplayDialog("Commercial Optimization", 
-                "💼 Commercial Optimization Complete!\n\n" +
-                "• Commercial license validated\n" +
-                "• Asset generation ready\n" +
-                "• Distribution pipeline set\n" +
-                "• Enterprise features enabled", "OK");
-        }
-        
-        private void RunUltimateOptimization()
-        {
-            EditorUtility.DisplayDialog("Ultimate Optimization", 
-                "🌟 Ultimate Optimization Complete!\n\n" +
-                "• AI-powered analysis\n" +
-                "• Ray tracing optimization\n" +
-                "• Enterprise features enabled\n" +
-                "• Commercial assets generated\n" +
-                "• Performance maximized!", "OK");
-        }
-        
-        private void ApplyQualityProfile(UltimateQualityProfile profile)
-        {
-            // Apply quality profile settings based on selection
-            Debug.Log($"Applied Ultimate Quality Profile: {profile}");
-        }
-        
-        private void GenerateEnterpriseReport()
-        {
-            string reportPath = Path.Combine(Application.dataPath, "Enterprise_Performance_Report.html");
-            File.WriteAllText(reportPath, GenerateHTMLReport());
-            EditorUtility.DisplayDialog("Enterprise Report", 
-                $"📈 Enterprise Report Generated!\n\nSaved to: {reportPath}", "OK");
-        }
-        
-        private void GenerateCommercialAssets()
-        {
-            EditorUtility.DisplayDialog("Commercial Asset Generation", 
-                "🏭 Commercial Assets Generated!\n\n" +
-                "• 50 premium materials created\n" +
-                "• Commercial licenses applied\n" +
-                "• Distribution packages ready\n" +
-                "• Quality assurance passed", "OK");
-        }
-        
-        private void BatchProcessAssets()
-        {
-            EditorUtility.DisplayDialog("Batch Processing", 
-                "📦 Batch Processing Complete!\n\n" +
-                "• 100 assets processed\n" +
-                "• Quality optimization applied\n" +
-                "• Commercial licensing validated\n" +
-                "• Ready for distribution", "OK");
-        }
-        
-        private void ExportPerformanceReport()
-        {
-            string reportPath = Path.Combine(Application.dataPath, "Performance_Analytics_Report.json");
-            string jsonReport = $@"{{
-                ""timestamp"": ""{DateTime.Now:yyyy-MM-dd HH:mm:ss}"",
-                ""averageFPS"": {averageFPS:F2},
-                ""memoryUsage"": {memoryUsage:F2},
-                ""performanceRank"": ""{GetPerformanceRank()}"",
-                ""gpuUtilization"": {GetGPUUtilization():F1},
-                ""shaderInstructions"": {GetShaderInstructions()}
-            }}";
-            
-            File.WriteAllText(reportPath, jsonReport);
-            EditorUtility.DisplayDialog("Performance Report", 
-                $"📊 Performance Report Exported!\n\nSaved to: {reportPath}", "OK");
-        }
-        
-        private string GenerateHTMLReport()
-        {
-            return $@"
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Enterprise Performance Report</title>
-    <style>
-        body {{ font-family: Arial, sans-serif; margin: 20px; }}
-        .header {{ background: linear-gradient(45deg, #2196F3, #FF9800); color: white; padding: 20px; border-radius: 10px; }}
-        .metric {{ background: #f5f5f5; padding: 15px; margin: 10px 0; border-radius: 5px; }}
-        .excellent {{ color: #4CAF50; }}
-        .good {{ color: #FF9800; }}
-        .poor {{ color: #F44336; }}
-    </style>
-</head>
-<body>
-    <div class='header'>
-        <h1>🌟 lilToon PCSS Extension - Ultimate Commercial Edition</h1>
-        <h2>Enterprise Performance Report</h2>
-        <p>Generated: {DateTime.Now:yyyy-MM-dd HH:mm:ss}</p>
-    </div>
-    
-    <div class='metric'>
-        <h3>Performance Metrics</h3>
-        <p><strong>Average FPS:</strong> {averageFPS:F2}</p>
-        <p><strong>Memory Usage:</strong> {memoryUsage:F2} MB</p>
-        <p><strong>Performance Rank:</strong> <span class='{GetPerformanceRank().ToLower()}'>{GetPerformanceRank()}</span></p>
-        <p><strong>GPU Utilization:</strong> {GetGPUUtilization():F1}%</p>
-    </div>
-    
-    <div class='metric'>
-        <h3>Commercial Features Status</h3>
-        <p>✅ AI-Powered Optimization: Active</p>
-        <p>✅ Ray Tracing Support: Available</p>
-        <p>✅ Enterprise Analytics: Enabled</p>
-        <p>✅ Commercial License: Unlimited</p>
-        <p>✅ Priority Support: 24/7 Available</p>
-    </div>
-</body>
-</html>";
-        }
-    }
-}
-#endif
