--- com.liltoon.pcss-extension-1.4.9-unity-test-clean/com.liltoon.pcss-extension-1.4.9-unity-test-clean/Runtime/lilToon.PCSS.Runtime.asmdef
@@ -1,40 +1,0 @@
-{
-    "name": "lilToon.PCSS.Runtime",
-    "rootNamespace": "lilToon.PCSS.Runtime",
-    "references": [
-        "Unity.TextMeshPro",
-        "Unity.RenderPipelines.Universal.Runtime",
-        "Unity.RenderPipelines.Core.Runtime"
-    ],
-    "optionalUnityReferences": [],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [
-        {
-            "name": "com.vrchat.avatars",
-            "expression": "",
-            "define": "VRCHAT_SDK_AVAILABLE"
-        },
-        {
-            "name": "nadena.dev.modular-avatar",
-            "expression": "",
-            "define": "MODULAR_AVATAR_AVAILABLE"
-        },
-        {
-            "name": "com.unity.render-pipelines.universal",
-            "expression": "",
-            "define": "UNITY_PIPELINE_URP"
-        },
-        {
-            "name": "com.unity.render-pipelines.high-definition",
-            "expression": "",
-            "define": "UNITY_PIPELINE_HDRP"
-        }
-    ],
-    "noEngineReferences": false
-}
